ecoScripts/service_now/Image/requirements.txt
kafka-python
pykafka
pyyaml
requests
pymongo
